third_party/WebKit/LayoutTests/fast/css/border-image-style-none-expected.txt

CONSOLE WARNING: Elements using the 'border-image' CSS property with no 'border-style' set should have no border, but currently do. Setting 'border-style' will be required in M51, around May 2016. See https://www.chromestatus.com/features/5542503914668032 for more details.
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is false
PASS window.internals.isUseCounted(document, BorderImageWithBorderStyleNone) is true
PASS successfullyParsed is true

TEST COMPLETE
